Validacion V4

+++ src/main/java/com/unab/banca/Controller/ClienteController.java
@@ -15,8 +15,10 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.validation.BindingResult;
 
 import com.unab.banca.Dto.ClienteDto;
 import com.unab.banca.Entity.Cliente;
@@ -24,6 +26,13 @@ import com.unab.banca.Entity.Message;
 import com.unab.banca.Service.ClienteService;
 import com.unab.banca.Utility.ConvertEntity;
 import com.unab.banca.Utility.Security.Hash;
+import com.unab.banca.Validation.Exception.InvalidDataException;
+import com.unab.banca.Validation.Exception.NoAuthorizeException;
+import com.unab.banca.Validation.Exception.UniqueException;
+
+import net.bytebuddy.asm.Advice.Return;
+
+import com.unab.banca.Validation.Entity.Error;
 
 @CrossOrigin(origins = "*")
 @RestController
@@ -39,14 +48,19 @@ public class ClienteController {
     ClienteDto clienteDto = new ClienteDto();
 
     @PostMapping("/create")
-    public ResponseEntity<Object> save(@Valid @RequestBody Cliente cliente) {
+    public ResponseEntity<Object> save(@RequestHeader String user, @RequestHeader String key,
+            @Valid @RequestBody Cliente cliente, BindingResult result) {
+        validarDatos(user, key, cliente.getNombre(), result);
         cliente.setClave(Hash.sha1(cliente.getClave()));
         return new ResponseEntity<>(convertEntity.convert(clienteService.save(cliente), clienteDto),
                 HttpStatus.CREATED);
     }
 
     @PutMapping("/update/{id}")
-    public ResponseEntity<Object> update(@Valid @RequestBody Cliente cliente, @PathVariable("id") String id) {
+    public ResponseEntity<Object> update(@PathVariable("id") String id, @RequestHeader String user,
+            @RequestHeader String key, @Valid @RequestBody Cliente cliente, BindingResult result) {
+
+        validarDatos(user, key, result);
         cliente.setIdCliente(id);
         cliente.setClave(Hash.sha1(cliente.getClave()));
         return new ResponseEntity<>(convertEntity.convert(clienteService.save(cliente), clienteDto), HttpStatus.OK);
@@ -66,7 +80,9 @@ public class ClienteController {
     @GetMapping("/list/{valor}")
     public ResponseEntity<Object> findByName(@PathVariable("valor") String valor) {
         if (clienteService.findByNombre(valor) == null) {
-            Message message = new Message(404, "usuario no encotrado con valor [" + valor + "]");
+            Message message = new Message();
+            message.setStatus(404);
+            message.setMessage("usuario no encotrado con valor [" + valor + "]");
             return new ResponseEntity<>(message, HttpStatus.NOT_FOUND);
         }
         return new ResponseEntity<>(clienteService.findByNombre(valor), HttpStatus.OK);
@@ -83,7 +99,41 @@ public class ClienteController {
     }
 
     @DeleteMapping("/delete/{id}")
-    public Object deleteById(@PathVariable("id") String id) {
+    public String deleteById(@PathVariable("id") String id, @RequestHeader String user, @RequestHeader String key) {
+        if (clienteService.logIn(user, Hash.sha1(key)) == 0) {
+
+            return "Acceso no autorizado";
+        }
         return clienteService.deleteById(id);
     }
+
+    private Boolean validarDatos(String user, String key, String nombre, BindingResult result) {
+
+        if (clienteService.logIn(user, Hash.sha1(key)) == 0) {
+            throw new NoAuthorizeException("Acceso No Autorizado", 
+            new Error("Campo nombre", "Acceso no Autorizado "));
+        }
+        if (result.hasErrors()) {
+            throw new InvalidDataException("Datos de entrada con errores", result);
+        }
+        if (clienteService.findByNombre(nombre) != null) {
+
+            throw new UniqueException("Datos de entrada con errores",
+                    new Error("Credenciales", "Ya existe un registro con el nombre " + nombre));
+        }
+        return true;
+    }
+
+    private Boolean validarDatos(String user, String key,  BindingResult result) {
+        if (clienteService.logIn(user, Hash.sha1(key)) == 0) {
+            throw new NoAuthorizeException("Acceso No Autorizado", 
+            new Error("Campo nombre", "Acceso no Autorizado "));
+        }
+        if (result.hasErrors()) {
+            throw new InvalidDataException("Datos de entrada con errores", result);
+        }
+
+        return true;
+    }
+
 }
